Refactor: extract FeedbackReplyAdapter and StatsAdapter
- Extracted FeedbackReplyAdapter and its DiffUtils logic from FeedbackDetailActivity into a standalone file.
- Extracted StatsAdapter and its ViewHolder from UserProfileFragment into a standalone file.
- Cached ContextCompat color lookups in StatsAdapter to improve onBindViewHolder performance.
- Removed unused imports and references in host files.
