+++ app/src/main/java/org/ole/planet/myplanet/ui/feedback/FeedbackDetailActivity.kt
@@ -26,8 +26,6 @@ import org.ole.planet.myplanet.databinding.RowFeedbackReplyBinding
 import org.ole.planet.myplanet.model.FeedbackReply
 import org.ole.planet.myplanet.model.RealmFeedback
 import org.ole.planet.myplanet.ui.dashboard.DashboardActivity
-import org.ole.planet.myplanet.ui.feedback.FeedbackDetailActivity.FeedbackReplyAdapter.ReplyViewHolder
-import org.ole.planet.myplanet.utils.DiffUtils
 import org.ole.planet.myplanet.utils.EdgeToEdgeUtils
 import org.ole.planet.myplanet.utils.LocaleUtils
 import org.ole.planet.myplanet.utils.TimeUtils.getFormattedDateWithTime
@@ -136,30 +134,5 @@ class FeedbackDetailActivity : AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         if (item.itemId == android.R.id.home) finish()
         return super.onOptionsItemSelected(item)
-    }
-
-    inner class FeedbackReplyAdapter(var context: Context) : ListAdapter<FeedbackReply, ReplyViewHolder>(DIFF_CALLBACK) {
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ReplyViewHolder {
-            val rowFeedbackReplyBinding = RowFeedbackReplyBinding.inflate(layoutInflater, parent, false)
-            return ReplyViewHolder(rowFeedbackReplyBinding)
-        }
-
-        override fun onBindViewHolder(holder: ReplyViewHolder, position: Int) {
-            val feedbackReply = getItem(position)
-            holder.binding.tvDate.text = feedbackReply.date.let {
-                getFormattedDateWithTime(it.toLong())
-            }
-            holder.binding.tvUser.text = feedbackReply.user
-            holder.binding.tvMessage.text = feedbackReply.message
-        }
-
-        inner class ReplyViewHolder(val binding: RowFeedbackReplyBinding) : RecyclerView.ViewHolder(binding.root)
-    }
-
-    companion object {
-        val DIFF_CALLBACK = DiffUtils.itemCallback<FeedbackReply>(
-            areItemsTheSame = { oldItem, newItem -> oldItem.date == newItem.date && oldItem.user == newItem.user },
-            areContentsTheSame = { oldItem, newItem -> oldItem == newItem }
-        )
     }
 }
